Extract low-risk routers: meta, callsign, config, weather
Move health check + Swagger docs into routes/meta.js, callsign lookup
(with HamQTH/callook helpers) into routes/callsign.js, config/feedback
into routes/config.js, and all weather endpoints (WU, NWS, OWM, radar,
clouds) into routes/weather.js. server.js reduced by another 733 lines.
